requirements.txt
Django==4.2.3
django-timezone-utils==0.15.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
celery[redis]==5.3.1
